Back-End com Python e Django/ex75/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
iniconfig==2.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.0
pytest-testdox==3.1.0
setuptools==76.0.0
sqlparse==0.5.3
tzdata==2025.2
